[Backend] Moved LocationRepository
