+++ backend/src/main/java/de/neuefische/flooooooooooorian/backend/repository/LocationRepository.java
@@ -1,4 +1,4 @@
-package de.neuefische.flooooooooooorian.backend;
+package de.neuefische.flooooooooooorian.backend.repository;
 
 import de.neuefische.flooooooooooorian.backend.model.Location;
 import org.springframework.data.repository.PagingAndSortingRepository;
